separete logic into hooks
